… radis/default_radis.json
… radis/lbl/broadening.py
… radis/spectrum/spectrum.py